+++ src/main/ts/ephox/alloy/behaviour/sandboxing/SandboxApis.ts
@@ -1,5 +1,5 @@
 import { Attr, Css, Element } from '@ephox/sugar';
-import { Option } from '@ephox/katamari';
+import { Option, Arr } from '@ephox/katamari';
 import { Positioning } from '../../api/behaviour/Positioning';
 import * as Attachment from '../../api/system/Attachment';
 import { AlloyComponent } from '../../api/component/ComponentApi';
@@ -89,7 +89,15 @@ const cloak = (sandbox: AlloyComponent, sConfig: SandboxingConfig, sState: Sandb
   store(sandbox, 'visibility', sConfig.cloakVisibilityAttr(), 'hidden');
 };
 
+const hasPosition = (element: Element) => Arr.exists(['top', 'left', 'right', 'bottom'], (pos) => Css.getRaw(element, pos).isSome());
+
 const decloak = (sandbox: AlloyComponent, sConfig: SandboxingConfig, sState: SandboxingState) => {
+  if (!hasPosition(sandbox.element())) {
+    // If a position value was not added to the sandbox during cloaking, remove it
+    // otherwise certain position values (absolute, relative) will impact the child that _was_ positioned
+    Css.remove(sandbox.element(), 'position');
+  }
+
   restore(sandbox, 'visibility', sConfig.cloakVisibilityAttr());
 };