+++ src/main/ts/ephox/alloy/behaviour/sliding/ActiveSliding.ts
@@ -1,14 +1,14 @@
 import { Objects } from '@ephox/boulder';
-import { Css } from '@ephox/sugar';
 import { TransitionEvent } from '@ephox/dom-globals';
+import { Css } from '@ephox/sugar';
 
+import { SugarEvent } from '../../alien/TypeDefinitions';
 import * as AlloyEvents from '../../api/events/AlloyEvents';
 import * as NativeEvents from '../../api/events/NativeEvents';
+import { SlidingConfig, SlidingState } from '../../behaviour/sliding/SlidingTypes';
 import * as DomModification from '../../dom/DomModification';
 import * as SlidingApis from './SlidingApis';
-import { SlidingConfig, SlidingState } from '../../behaviour/sliding/SlidingTypes';
-import { EventFormat, SimulatedEvent } from '../../events/SimulatedEvent';
-import { SugarEvent } from '../../alien/TypeDefinitions';
+import { getAnimationRoot } from './SlidingUtils';
 
 const exhibit = (base: { }, slideConfig: SlidingConfig/*, slideState */): { } => {
   const expanded = slideConfig.expanded();
@@ -25,9 +25,10 @@ const exhibit = (base: { }, slideConfig: SlidingConfig/*, slideState */): { } =>
 const events = (slideConfig: SlidingConfig, slideState: SlidingState): AlloyEvents.AlloyEventRecord => {
   return AlloyEvents.derive([
     AlloyEvents.run<SugarEvent>(NativeEvents.transitionend(), (component, simulatedEvent) => {
+      const aniRoot = getAnimationRoot(component, slideConfig).dom();
       const raw = simulatedEvent.event().raw() as TransitionEvent;
       // This will fire for all transitions, we're only interested in the dimension completion
-      if (raw.propertyName === slideConfig.dimension().property()) {
+      if (raw.target === aniRoot && raw.propertyName === slideConfig.dimension().property()) {
         SlidingApis.disableTransitions(component, slideConfig); // disable transitions immediately (Safari animates the dimension removal below)
         if (slideState.isExpanded()) { Css.remove(component.element(), slideConfig.dimension().property()); } // when showing, remove the dimension so it is responsive
         const notify = slideState.isExpanded() ? slideConfig.onGrown() : slideConfig.onShrunk();