+++ src/test/ts/browser/behaviour/CouplingTest.ts
@@ -24,7 +24,7 @@ UnitTest.asynctest('CouplingTest', (success, failure) => {
                   dom: {
                     tag: 'button'
                   },
-                  action: store.adder('clicked on coupled button of: ' + Attr.get(primary.element(), Tagger.attribute()))
+                  action: store.adder('clicked on coupled button of: ' + Tagger.read(primary.element()).getOr('No UID'))
                 });
               }
             }