+++ src/test/ts/browser/ui/slotcontainer/SlotContainerTest.ts
@@ -1,11 +1,11 @@
 import { ApproxStructure, Assertions, Chain, Step } from '@ephox/agar';
 import { UnitTest } from '@ephox/bedrock';
+import { Result } from '@ephox/katamari';
 import * as GuiFactory from 'ephox/alloy/api/component/GuiFactory';
 import { Button } from 'ephox/alloy/api/ui/Button';
 import { Input } from 'ephox/alloy/api/ui/Input';
 import { SlotContainer } from 'ephox/alloy/api/ui/SlotContainer';
 import * as GuiSetup from 'ephox/alloy/test/GuiSetup';
-import { Result } from '@ephox/katamari';
 
 UnitTest.asynctest('SlotContainerTest', (success, failure) => {
 
@@ -108,12 +108,39 @@ UnitTest.asynctest('SlotContainerTest', (success, failure) => {
       });
     });
 
+    const cSlotShowing = (slot: string) => Chain.mapper(() => {
+      return SlotContainer.isShowing(component, slot);
+    });
+
+    const sAssertSlotShowing = (slot: string, expectedShowing: boolean) => (label: string) => Chain.asStep({}, [
+      cSlotShowing(slot),
+      Assertions.cAssertEq('SlotContainer.isShowing(_,"' + slot + '") !== ' + expectedShowing + ' - ' + label, expectedShowing)
+    ]);
+
+    const sAssertButtonSlotShowing = sAssertSlotShowing('buttonB', true);
+    const sAssertButtonSlotHidden = sAssertSlotShowing('buttonB', false);
+    const sAssertInputSlotShowing = sAssertSlotShowing('inputA', true);
+    const sAssertInputSlotHidden = sAssertSlotShowing('inputA', false);
+
+    const cGetSlotNames = () => Chain.mapper(() => {
+      return SlotContainer.getSlotNames(component);
+    });
+
+    const sAssertGetSlotNames = (label: string, expectedSlots: string[]) => Chain.asStep({}, [
+      cGetSlotNames(),
+      Assertions.cAssertEq(label, expectedSlots)
+    ]);
+
     const sShowSlot = (slot: string) => Step.sync(() => {
       SlotContainer.showSlot(component, slot);
     });
 
     const sHideSlot = (slot: string) => Step.sync(() => {
       SlotContainer.hideSlot(component, slot);
+    });
+
+    const sHideAllSlots = () => Step.sync(() => {
+      SlotContainer.hideAllSlots(component);
     });
 
     return [
@@ -141,18 +168,34 @@ UnitTest.asynctest('SlotContainerTest', (success, failure) => {
 
       sAssertButtonShowing('button: Before any APIs are called'),
       sAssertInputShowing('input: Before any APIs are called'),
+      sAssertInputSlotShowing('before any APIs are called'),
+      sAssertButtonSlotShowing('before any APIs are called'),
       sHideSlot('buttonB'),
       sAssertButtonHidden('button: After SlotContainer.hideSlot(_, button)'),
+      sAssertButtonSlotHidden('After SlotContainer.hideSlot(_, button)'),
       sAssertInputShowing('input: After SlotContainer.hideSlot(_, button)'),
+      sAssertInputSlotShowing('After SlotContainer.hideSlot(_, button)'),
       sShowSlot('buttonB'),
       sAssertButtonShowing('button: After SlotContainer.showSlot(_, button)'),
+      sAssertButtonSlotShowing('After SlotContainer.showSlot(_, button)'),
       sAssertInputShowing('input: After SlotContainer.showSlot(_, button)'),
+      sAssertInputSlotShowing('After SlotContainer.showSlot(_, button)'),
       sHideSlot('inputA'),
       sAssertButtonShowing('button: After SlotContainer.hideSlot(_, input)'),
+      sAssertButtonSlotShowing('After SlotContainer.hideSlot(_, input)'),
       sAssertInputHidden('input: After SlotContainer.hideSlot(_, input)'),
+      sAssertInputSlotHidden('After SlotContainer.hideSlot(_, input)'),
       sShowSlot('inputA'),
       sAssertButtonShowing('button: After SlotContainer.showSlot(_, input)'),
       sAssertInputShowing('input: After SlotContainer.showSlot(_, input)'),
+      sAssertInputSlotShowing('After SlotContainer.showSlot(_, input)'),
+      sAssertButtonSlotShowing('After SlotContainer.showSlot(_, input)'),
+      sHideAllSlots(),
+      sAssertButtonHidden('button: After SlotContainer.hideAllSlots(_)'),
+      sAssertInputHidden('input: After SlotContainer.hideAllSlots(_)'),
+      sAssertInputSlotHidden('After SlotContainer.hideAllSlots(_)'),
+      sAssertButtonSlotHidden('After SlotContainer.hideAllSlots(_)'),
+      sAssertGetSlotNames('checking the list slots', ['inputA', 'buttonB']),
     ];
   }, () => { success(); }, failure);
 });